move: userRepository를 business 계층으로 이동
