+++ src/main/java/com/newzet/api/user/repository/repository/UserRepositoryImpl.java
@@ -3,6 +3,7 @@ package com.newzet.api.user.repository.repository;
 import org.springframework.stereotype.Repository;
 
 import com.newzet.api.user.business.dto.UserEntityDto;
+import com.newzet.api.user.business.service.UserRepository;
 import com.newzet.api.user.repository.entity.UserEntity;
 import com.newzet.api.user.repository.exception.NoUserException;
